+++ trunk/src/wink/wink-rand32.hpp
@@ -8,9 +8,9 @@
 #ifndef WINK_RAND32_INCLUDED
 #define WINK_RAND32_INCLUDED 1
 
-#include <stddef.h>
-#include <stdint.h>
-#include <assert.h>
+#include "wink-os.hpp"
+#include <cassert>
+#include <cstdlib>
 
 namespace wink 
 {
@@ -57,7 +57,7 @@ namespace wink
         }
         
         //! convert to 0:1 exclusive
-        static inline double to_float( uint32_t u ) throw()
+        static inline float to_float( uint32_t u ) throw()
         {
             return (0.5f+float(u))/4294967296.0f;
         }
